+++ megatron/core/datasets/gpt_dataset.py
@@ -414,7 +414,7 @@ class GPTDataset(MegatronDataset):
             else:
                 drop_last_partial_sequence = True
 
-            assert document_index.dtype == numpy.int32
+            assert document_index.dtype == numpy.int64
             assert self.dataset.sequence_lengths.dtype == numpy.int32
             sample_index = helpers.build_sample_idx(
                 self.dataset.sequence_lengths,
@@ -556,7 +556,7 @@ def _build_document_index(
         document_index = numpy.mgrid[0:num_epochs, 0 : len(documents)][1]
         document_index[:] = documents
         document_index = document_index.reshape(-1)
-        document_index = document_index.astype(numpy.int32)
+        document_index = document_index.astype(numpy.int64)
         numpy_random_state.shuffle(document_index)
         return document_index
